Renamed aliasDirectory to aliasMappings to more accurately describe the feature, and updated the README.
