+++ test/fixtures/aliasMappings/entry.js
