+++ test/fixtures/aliasMappings/foo/bar/foobar.js
