+++ test/fixtures/aliasMappings/foo/foo.js
