+++ test/fixtures/aliasMappings/root.js
